% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{assessment}
\alias{assessment}
\title{sboost Assessment Function}
\usage{
assessment(features, outcomes, classifier)
}
\arguments{
\item{outcomes}{outcomes corresponding to the features}

\item{classifier}{must be output from sboost}

\item{feature}{feature set data.frame}
}
\value{
Assessment after each stump in the classifier.
}
\description{
Assesses how well a classifier fits the data.
}
\details{
See sboost documentation for more details.
}
\keyword{accuracy}
\keyword{assess,}
\keyword{assessment,}
\keyword{f1,}
