% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{classifier_predictions}
\alias{classifier_predictions}
\title{predictions makes predictions for a feature set}
\usage{
classifier_predictions(features, outcome_possibilities, classifier)
}
\arguments{
\item{features}{feature set data.frame}

\item{outcome_possibilities}{possible values of outcomes in a vector}

\item{classifier}{classifier output from sboost}
}
\value{
predictions in the form of a vector
}
\description{
predictions makes predictions for a feature set
}
\keyword{boost,}
\keyword{classifier}
\keyword{stump,}
