% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malware}
\alias{malware}
\title{Malware System Calls}
\format{A data frame with 7597 rows and 361 variables: \emph{outcomes} 1 if malware, 0 if not.
\emph{X1... X360} system calls.}
\source{
\url{https://zenodo.org/record/154737#.WtoA1IjwaUl}
}
\usage{
malware
}
\description{
System call data for apps identified as malware and not malware.
}
\details{
Experimental data generated in this research paper:

M. Dimjašević, S. Atzeni, I. Ugrina, and Z. Rakamarić,
"Evaluation of Android Malware Detection Based on System Calls,"
in Proceedings of the International Workshop on Security and Privacy Analytics (IWSPA), 2016.

Data used for kaggle competition: \url{https://www.kaggle.com/c/ml-fall2016-android-malware}
}
\keyword{datasets}
