% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions}
\alias{predictions}
\title{Make predictions for a feature set based on an sboost classifier.}
\usage{
predictions(features, outcome_possibilities, classifier)
}
\arguments{
\item{features}{feature set data.frame}

\item{outcome_possibilities}{possible values of outcomes in a vector}

\item{classifier}{classifier output from sboost}
}
\value{
predictions in the form of a vector
}
\description{
See sboost documentation for more details.
}
\keyword{predict}
\keyword{predictions,}
