% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{sboost Assessment Function}
\usage{
assess(object, features, outcomes, include_scores = FALSE)
}
\arguments{
\item{object}{\emph{sboost_classifier} S3 object output from sboost.}

\item{features}{feature set data.frame.}

\item{outcomes}{outcomes corresponding to the features.}

\item{include_scores}{if true feature_scores are included in output.}
}
\description{
Assesses how well an sboost classifier classifies the data.
}
\examples{
# malware
malware_classifier <- sboost(malware[-1], malware[1], iterations = 10, positive = 1)
assess(malware_classifier, malware[-1], malware[1])

# mushrooms
mushroom_classifier <- sboost(mushrooms[-1], mushrooms[1], iterations = 10, positive = "p")
assess(mushroom_classifier, mushrooms[-1], mushrooms[1])
}
\seealso{
\code{\link{sboost}} documentation.
}
