% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_classifier.R
\name{assess_classifier}
\alias{assess_classifier}
\title{assess_model}
\usage{
assess_classifier(features, outcomes, classifier)
}
\arguments{
\item{outcomes}{index corresponding to the features}

\item{classifier}{must be output from sboost}

\item{feature}{feature set}
}
\value{
Assessment after each additional line in the model
}
\description{
This function finds determines how well a model fits a given dataset
}
\keyword{accuracy}
\keyword{assess,}
\keyword{assessment,}
\keyword{f1,}
