% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{assessment}
\alias{assessment}
\title{sboost Assessment Function}
\usage{
assessment(features, outcomes, classifier, positive = NULL)
}
\arguments{
\item{outcomes}{outcomes corresponding to the features}

\item{classifier}{must be output from sboost}

\item{positive}{is the positive outcome to test for; if NULL, the first in alphebetacal order will be chosen}

\item{feature}{feature set data.frame}
}
\value{
Assessment after each stump in the classifier.
}
\description{
Assesses how well a classifier fits the data.
}
\details{
See sboost documentation for more details.
}
\examples{
# malware
malware_classifier <- sboost(malware[-1], malware[1], iterations = 10)
assessment(malware[-1], malware[1], malware_classifier, positive = 1)

# mushrooms
mushroom_classifier <- sboost(mushrooms[-1], mushrooms[1], iterations = 10)
assessment(mushrooms[-1], mushrooms[1], mushroom_classifier, positive = "p")
}
\keyword{accuracy}
\keyword{assess,}
\keyword{assessment,}
\keyword{f1,}
