% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions}
\alias{predictions}
\title{Make predictions for a feature set based on an sboost classifier.}
\usage{
predictions(features, classifier, scores = FALSE)
}
\arguments{
\item{features}{feature set data.frame.}

\item{classifier}{\emph{sboost_classifier} S3 object output from sboost.}

\item{scores}{if true, raw scores generated; if false, predictions are generated.}
}
\value{
Predictions in the form of a vector, or scores in the form of a vector.
  The index of the vector aligns the predictions or scores with the rows of
  the features. Scores represent the sum of all votes for the positive outcome
  minus the sum of all votes for the negative outcome.
}
\description{
Make predictions for a feature set based on an sboost classifier.
}
\examples{
# malware
malware_classifier <- sboost(malware[-1], malware[1], iterations = 10, positive = 1)
predictions(malware[-1], malware_classifier, scores = TRUE)
predictions(malware[-1], malware_classifier)

# mushrooms
mushroom_classifier <- sboost(mushrooms[-1], mushrooms[1], iterations = 10, positive = "p")
predictions(mushrooms[-1], mushroom_classifier, scores = TRUE)
predictions(mushrooms[-1], mushroom_classifier)
}
\seealso{
\code{\link{sboost}} documentation.
}
