% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions}
\alias{predictions}
\title{Make predictions for a feature set based on an sboost classifier.}
\usage{
predictions(features, classifier)
}
\arguments{
\item{features}{feature set data.frame}

\item{classifier}{classifier output from sboost}

\item{outcome_possibilities}{possible values of outcomes in a vector}
}
\value{
predictions in the form of a vector
}
\description{
See sboost documentation for more details.
}
\examples{
# malware
malware_classifier <- sboost(malware[-1], malware[1], iterations = 10)
predictions(malware[-1], malware[1], malware_classifier)

# mushrooms
mushroom_classifier <- sboost(mushrooms[-1], mushrooms[1], iterations = 10)
predictions(mushrooms[-1], mushrooms[1], mushroom_classifier)
}
\keyword{predict}
\keyword{predictions,}
