% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sboost.R
\name{sboost}
\alias{sboost}
\title{sboost learning algorithm}
\usage{
sboost(features, outcomes, iterations = 1)
}
\arguments{
\item{outcomes}{outcomes corresponding to the features}

\item{iterations}{is the number of boosts}

\item{feature}{feature set data.frame}
}
\value{
classifier
}
\description{
This algorithm is a machine learning algorithm using
adaboost on decision stumps.
}
\keyword{boost}
\keyword{stump,}
