% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sboost.R
\name{sboost}
\alias{sboost}
\title{sboost Learning Algorithm}
\usage{
sboost(features, outcomes, iterations = 1)
}
\arguments{
\item{features}{feature set data.frame}

\item{outcomes}{outcomes corresponding to the features}

\item{iterations}{number of boosts}
}
\value{
Classifier where each row is a stump.
}
\description{
A machine learning algorithm using adaboost on decision stumps.
}
\details{
Factors and characters are treated as categorical features.
Missing values are treated as their own category for categorical,
and divided randomly on the best split for numeric.
}
\keyword{adaboost,}
\keyword{boost,}
\keyword{classifier,}
\keyword{decision}
\keyword{stump}
\keyword{stump,}
