% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sboost.R
\name{sboost}
\alias{sboost}
\title{sboost Learning Algorithm}
\usage{
sboost(features, outcomes, iterations = 1, positive = NULL)
}
\arguments{
\item{features}{feature set data.frame}

\item{outcomes}{outcomes corresponding to the features}

\item{iterations}{number of boosts}

\item{positive}{the positive outcome to test for; if NULL, the first outcome in alphebetacal order will be chosen}
}
\value{
A list:
}
\description{
A machine learning algorithm using adaboost on decision stumps.
}
\details{
Factors and characters are treated as categorical features.
Missing values are treated as their own category for categorical,
and divided randomly on the best split for numeric.
}
\examples{
# malware
malware_classifier <- sboost(malware[-1], malware[1], iterations = 10)
malware_classifier
malware_classifier$classifier

# mushrooms
mushroom_classifier <- sboost(mushrooms[-1], mushrooms[1], iterations = 10)
mushroom_classifier
mushroom_classifier$classifier
}
\keyword{adaboost,}
\keyword{boost,}
\keyword{classifier,}
\keyword{decision}
\keyword{stump}
\keyword{stump,}
