% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{sboost Validation Function}
\usage{
validate(features, outcomes, iterations = 1, k_fold = 6)
}
\arguments{
\item{outcomes}{outcomes corresponding to the features}

\item{iterations}{number of boosts}

\item{k_fold}{number of cross-validation subsets}

\item{feature}{feature set data.frame}
}
\value{
Statistics for the test set and the training set.
}
\description{
A k-fold cross validation algorithm for sboost.
}
\details{
See sboost documentation for more details.
}
\keyword{validate}
\keyword{validation,}
