% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{sboost Validation Function}
\usage{
validate(features, outcomes, iterations = 1, k_fold = 6,
  positive = NULL)
}
\arguments{
\item{features}{feature set data.frame}

\item{outcomes}{outcomes corresponding to the features}

\item{iterations}{number of boosts}

\item{k_fold}{number of cross-validation subsets}

\item{positive}{is the positive outcome to test for; if NULL, the first in alphebetacal order will be chosen}
}
\value{
Statistics for the test set and the training set.
}
\description{
A k-fold cross validation algorithm for sboost.
}
\details{
See sboost documentation for more details.
}
\examples{
# malware
validate(malware[-1], malware[1], iterations = 10, k_fold = 4, positive = 1)

# mushrooms
validate(mushrooms[-1], mushrooms[1], iterations = 10, k_fold = 4, positive = "p")
}
\keyword{validate}
\keyword{validation,}
