% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{sboost learning algorithm validation}
\usage{
validate(features, outcomes, iterations = 1, k_fold = 6)
}
\arguments{
\item{outcomes}{outcomes corresponding to the features}

\item{iterations}{is the number of boosts}

\item{k_fold}{is the number of cross-validation subsets}

\item{feature}{feature set data.frame}
}
\value{
classifier
}
\description{
This is a k-fold cross validation algorithm for sboost.
}
\keyword{boost,}
\keyword{stump,}
\keyword{validation}
